backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.1
sqlmodel==0.0.22
alembic==1.13.2
pydantic==2.9.2
pydantic-core==2.23.4
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
pdfplumber==0.7.6
python-docx==0.8.11
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.1
pytesseract==0.3.10
